refactor: add scheme toggle to header in debug mode
